fix: surface non-zero exit code for warnings
Currently if clj-kondo detects warnings, it returns an exit code of `2`;
however, the `clojure-lint-action` script swallows the `2` and exits
with `0`. In this situation, it would be great to return a non-zero
exit code upstream so that the GH action reports a failure.

If consumers of the GH action only want the check to fail on errors,
they can provide the `--fail-level error` arg:

`clj-kondo --lint src --fail-level error`

This will only return a 0 if there are no errors.